+++ lib/services/service_brokers/v2.rb
@@ -7,6 +7,7 @@ require 'services/service_brokers/v2/catalog_plan'
 require 'services/service_brokers/v2/catalog_schemas'
 require 'services/service_brokers/v2/schema'
 require 'services/service_brokers/v2/http_client'
+require 'services/service_brokers/v2/http_response'
 require 'services/service_brokers/v2/client'
 require 'services/service_brokers/v2/orphan_mitigator'
 require 'services/service_brokers/v2/response_parser'
