+++ spec/acceptance/broker_api_compatibility/broker_api_v2.15_spec.rb
@@ -28,5 +28,44 @@ RSpec.describe 'Service Broker API integration' do
         end
       end
     end
+
+    describe 'platform delays polling to last_operation based on Retry-After header' do
+      let(:default_poll_interval) { VCAP::CloudController::Config.config.get(:broker_client_default_async_poll_interval_seconds) }
+      let(:retry_after_interval) { default_poll_interval * 4 }
+
+      context 'when provisioning a service instance' do
+        before do
+          setup_broker(default_catalog)
+          @broker = VCAP::CloudController::ServiceBroker.find guid: @broker_guid
+        end
+
+        it 'should schedule a delayed job with correct run_at time' do
+          stub_async_last_operation(body: { state: 'in progress' }, headers: { 'Retry-After': retry_after_interval })
+          async_provision_service
+
+          expect(
+            a_request(:put, provision_url_for_broker(@broker, accepts_incomplete: true))
+          ).to have_been_made
+
+          service_instance = VCAP::CloudController::ManagedServiceInstance.find(guid: @service_instance_guid)
+
+          Timecop.freeze(Time.now) do
+            # Initial call is made immediately after the job is scheduled
+            Delayed::Worker.new.work_off
+            expect(a_request(:get, %r{#{service_instance_url(service_instance)}/last_operation})).to have_been_made
+
+            # Check a call on the default interval has not been made, i.e. still have the initial call to last_operation
+            Timecop.travel(default_poll_interval.seconds)
+            Delayed::Worker.new.work_off
+            expect(a_request(:get, %r{#{service_instance_url(service_instance)}/last_operation})).to have_been_made.once
+
+            # Check a new call has been made at the Retry-After interval
+            Timecop.travel(retry_after_interval.seconds)
+            Delayed::Worker.new.work_off
+            expect(a_request(:get, %r{#{service_instance_url(service_instance)}/last_operation})).to have_been_made.twice
+          end
+        end
+      end
+    end
   end
 end
